Updated Simulated Annealing Codes
This update allows for directory independence. That is: source codes, instance files, and output files can all be held in different, independent locations. 

All output is now formatted in a useful way. Instead of returning the byte data type, properly formatted tuples of only the most important information is returned.

Dependence on the creation of batch files was rendered obsolete and thus removed.

The tutorial was updated to reflect these changes. The doc-string was updated to include example usage of the function.